Source code here
Github desktop is mad, there is to many files, and can't do zip files upload directly on the website either they are still to big.
Something must have changed, was never a problem before to upload. Pretty sure I could upload a unity project by drag and drop it on the website, at least.
But the source code is the important part which is present, so hopefully that is good enough for the Ludum Dare.